pyproject.toml
[tool.poetry]
name = "tradeai"
version = "1.0.0"
description = "AI-powered trade signal platform"
authors = ["TradeAI"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.0"}
asyncpg = ">=0.30.0"
alembic = ">=1.14.0"
pydantic-settings = ">=2.7.0"
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
slowapi = ">=0.1.9"
structlog = ">=24.4.0"
tenacity = ">=9.0.0"
httpx = ">=0.28.0"
anthropic = ">=0.43.0"
yfinance = ">=0.2.50"
pandas = ">=2.2.0"
ta = ">=0.11.0"
apscheduler = ">=3.10.0"
python-multipart = ">=0.0.20"
greenlet = "^3.3.2"
pydantic = {extras = ["email"], version = "^2.12.5"}
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0"
pytest-asyncio = ">=0.24.0"
ruff = ">=0.8.0"
mypy = ">=1.13.0"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
